done done

+++ src/components/Lock/MyItem.js
@@ -3,7 +3,6 @@ import { useAuthState } from "react-firebase-hooks/auth";
 import auth from "../firebase.init";
 import { MdDeleteForever } from "react-icons/md";
 import { useNavigate } from "react-router-dom";
-import axios from "axios";
 
 const MyItem = () => {
   const navigate = useNavigate();
@@ -19,16 +18,19 @@ const MyItem = () => {
   console.log(products);
 
   const deleteHandeler = (id) => {
-    fetch(`https://intense-stream-06695.herokuapp.com/user/${id}`, {
-      method: "delete",
-    })
-      .then((response) => response.json())
-      .then((data) => {
-        if (data.deletedCount > 0) {
-          navigate("/");
-          alert("Delete success");
-        }
-      });
+    const confirm = window.confirm("Are You Sure");
+    if (confirm) {
+      fetch(`https://intense-stream-06695.herokuapp.com/user/${id}`, {
+        method: "delete",
+      })
+        .then((response) => response.json())
+        .then((data) => {
+          if (data.deletedCount > 0) {
+            navigate("/");
+            alert("Delete success");
+          }
+        });
+    }
   };
   return (
     <div>
